Fix comments and update naming file
