+++ big-practice/src/constants/url.js
@@ -1,0 +1,1 @@
+export const BASE_URL = 'https://64a67a49096b3f0fcc7fda51.mockapi.io/ngando/';
