+++ big-practice/src/index.html
@@ -41,7 +41,7 @@
         </tr>
       </thead>
 
-      <tbody class="table-main">
+      <tbody class="table-customer-body">
       </tbody>
     </table>
 
@@ -110,7 +110,7 @@
     <!-- end modal update customer -->
 
     <!-- start modal delete -->
-    <section class="del-confirm-modal">
+    <section class="modal-customer">
       <div class="modal-header">
         <h2 class="form-title">Do you want to remove this customer?</h2>
         <img src="./assets/images/cancel.png" alt="Icon cancel update customer form">
