+++ big-practice/src/scripts/app.js
@@ -1,6 +1,6 @@
-import CustomerController from "./controllers/customer.controller";
-import CustomerView from "./views/customer.view";
-import CustomerService from './services/customer.service';
+import CustomerController from "./controllers/customer-controller";
+import CustomerView from "./views/customer-view";
+import CustomerService from './services/customer-service';
 
 export class App {
     constructor() {}
@@ -8,7 +8,7 @@ export class App {
     startApp() {
         // Init the customer controller
         const customerController = new CustomerController(new CustomerView(), new CustomerService());
-       
+
         customerController.init();
     }
 }
