+++ big-practice/src/scripts/controllers/customer-controller.js
@@ -4,7 +4,7 @@ class CustomerController {
     this.service = service;
   }
 
-  init() {
+  init = () => {
     this.handleRenderTable();
   }
