+++ big-practice/src/scripts/views/customer-view.js
@@ -2,18 +2,7 @@ import imgDot from '../../assets/images/dot.png';
 
 class CustomerView {
   constructor() {
-    this.table = document.querySelector('.table-main');
-    this.iconModal = document.querySelector('.icon-modal');
-    this.modalCustomer = document.querySelector('.modal-customer');
-    this.iconCancel = document.querySelector('.icon-cancel');
-
-    this.iconModal.addEventListener("click", () => {
-      this.modalCustomer.style.display = "block";
-    });
-
-    this.iconCancel.addEventListener("click", () => {
-      this.modalCustomer.style.display = "none";
-    });
+    this.table = document.querySelector('.table-customer .table-customer-body');
   }
 
   renderData(list) {
