+++ big-practice/src/styles/components/all.css
@@ -2,7 +2,7 @@
 @import url("./common/all.css");
 
 /* Components styles */
-@import url("./del-confirm-modal.css");
+@import url("./status-action.css");
 @import url("./modal-customer.css");
 @import url("./table-customer.css");
 @import url("./container.css");
