+++ big-practice/src/styles/components/table-customer.css
@@ -61,7 +61,7 @@
 }
 
 .dropdown .select-dropdown {
-  color: #7E7E7E;
+  color: #7e7e7e;
   font-size: 12px;
   font-family: var(--primary-font);
   font-style: normal;
@@ -128,7 +128,7 @@
   padding: 4px 12px;
   border-radius: 5px;
   border-radius: 4px;
-  border: 1px solid #00B087;
+  border: 1px solid #00b087;
   background: rgba(22, 192, 152, 0.38);
   color:#008767;
   font-size: 14px;
@@ -153,8 +153,8 @@
   width: 26px;
   height: 24px;
   border-radius: 4px;
-  border: 1px solid #EEE;
-  background: #F5F5F5;
+  border: 1px solid #eee;
+  background: #f5f5f5;
   margin-right: 10px;
-  color: #404B52;
+  color: #404b52;
 }
